requirements.txt
pyaudiowpatch>=0.2.12.8
openai>=1.0.0
anthropic>=0.18.0
winotify>=1.1.0
pystray>=0.19.0
Pillow>=10.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
pydub>=0.25.1
